feat: 구독 카테고리 목록 수정을 위한 모달 컴포넌트 구현 및 적용
